FDG-7840 refactored column formatting
