+++ src/components/data-table/data-table.tsx
@@ -1,14 +1,5 @@
 import React, { FunctionComponent, useEffect, useState } from 'react';
-import {
-  ColumnDef,
-  getCoreRowModel,
-  getPaginationRowModel,
-  getSortedRowModel,
-  useReactTable,
-  getFilteredRowModel,
-  SortingState,
-} from '@tanstack/react-table';
-import { currencyFormatter, numberFormatter } from '../../helpers/text-format/text-format';
+import { getCoreRowModel, getPaginationRowModel, getSortedRowModel, useReactTable, getFilteredRowModel, SortingState } from '@tanstack/react-table';
 import DataTableFooter from './data-table-footer/data-table-footer';
 
 import StickyTable from 'react-sticky-table-thead';
@@ -23,7 +14,7 @@ import {
 import DataTableHeader from './data-table-header/data-table-header';
 import DataTableColumnSelector from './column-select/data-table-column-selector';
 import DataTableBody from './data-table-body/data-table-body';
-import moment from 'moment';
+import { columnsConstructor } from './data-table-helper';
 
 type DataTableProps = {
   // defaultSelectedColumns will be null unless the dataset has default columns specified in the dataset config
@@ -43,16 +34,6 @@ type DataTableProps = {
   setFiltersActive: (value: boolean) => void;
 };
 
-const customFormat = (stringValue, decimalPlaces) => {
-  // if block is to show "-$123,123.23" instead of "$-123,123.23"
-  const absVal = Math.abs(stringValue);
-  let returnString = '$' + absVal.toFixed(decimalPlaces).replace(/(\d)(?=(\d{3})+(?!\d))/g, '$1,');
-  if (Number(stringValue) < 0) {
-    returnString = '-' + returnString;
-  }
-  return returnString;
-};
-
 const DataTable: FunctionComponent<DataTableProps> = ({
   rawData,
   defaultSelectedColumns,
@@ -69,83 +50,7 @@ const DataTable: FunctionComponent<DataTableProps> = ({
   pageSize,
   setFiltersActive,
 }) => {
-  console.log(rawData);
-  const allColumns = rawData.meta
-    ? Object.entries(rawData.meta.labels).map(([field, label]) => {
-        if (field === 'record_date') {
-          return {
-            accessorKey: field,
-            header: label,
-            filterFn: 'equalsString',
-            cell: ({ getValue }) => {
-              return moment(getValue()).format('M/D/YYYY');
-            },
-          } as ColumnDef<any, any>;
-        } else if (rawData.meta.dataTypes[field] === 'DATE') {
-          return {
-            accessorKey: field,
-            header: label,
-            cell: ({ getValue }) => {
-              return moment(getValue()).format('M/D/YYYY');
-            },
-          } as ColumnDef<any, any>;
-        } else if (rawData.meta.dataTypes[field] === 'NUMBER') {
-          return {
-            accessorKey: field,
-            header: label,
-            cell: ({ getValue }) => {
-              return numberFormatter.format(getValue());
-            },
-          } as ColumnDef<any, any>;
-        } else if (rawData.meta.dataTypes[field] === 'PERCENTAGE') {
-          return {
-            accessorKey: field,
-            header: label,
-            cell: ({ getValue }) => {
-              return `${getValue()}%`;
-            },
-          } as ColumnDef<any, any>;
-        } else if (rawData.meta.dataTypes[field] === 'SMALL_FRACTION') {
-          return {
-            accessorKey: field,
-            header: label,
-            cell: ({ getValue }) => {
-              return new Intl.NumberFormat('en-US', { maximumSignificantDigits: 5 }).format(getValue());
-            },
-          } as ColumnDef<any, any>;
-        } else if (rawData.meta.dataTypes[field] === 'CURRENCY') {
-          return {
-            accessorKey: field,
-            header: label,
-            cell: ({ getValue }) => {
-              return currencyFormatter.format(getValue());
-            },
-          } as ColumnDef<any, any>;
-        } else if (rawData.meta.dataTypes[field].includes('CURRENCY') && /\d/.test(rawData.meta.dataTypes[field].split('CURRENCY')[1])) {
-          const decimalPlaces = parseInt(rawData.meta.dataTypes[field].split('CURRENCY')[1]);
-          return {
-            accessorKey: field,
-            header: label,
-            cell: ({ getValue }) => {
-              return customFormat(getValue(), decimalPlaces);
-            },
-          } as ColumnDef<any, any>;
-        } else if (rawData.meta.dataTypes[field] === 'STRING') {
-          return {
-            accessorKey: field,
-            header: label,
-            cell: ({ getValue }) => {
-              if (getValue().includes('%')) {
-                return getValue().replace(/-/g, '\u2011');
-              } else {
-                return getValue();
-              }
-            },
-          } as ColumnDef<any, any>;
-        }
-        return { accessorKey: field, header: label } as ColumnDef<any, any>;
-      })
-    : [];
+  const allColumns = columnsConstructor(rawData);
   const data = rawData.data;
 
   if (hasPublishedReports && !hideCellLinks) {
